requirements.txt
flask>=3.0
gunicorn>=22.0
yfinance>=0.2
pandas>=2.0
numpy>=1.24
requests>=2.28
lxml>=4.9
google-genai>=1.0
